+++ setup.cfg
@@ -34,9 +34,11 @@ include_package_data = True
 
 [options.package_data]
 * = LICENSE.txt, CHANGELOG.md
+
 [options.extras_require]
 test =
     pytest
+
 [options.entry_points]
 console_scripts =
-    buildH = buildh:main
+    buildH = buildh.cli:main
